requirements.txt

torch
torchvision
torchaudio

pytorch-lightning
wandb

opencv-python
numpy
Pillow
tqdm

segmentation-models-pytorch

albumentations
albucore

warmup-scheduler
scikit-image

einops
transformers
pandas
